refactor: split chat components by feature
Refs #27
